Revert "pull request fixes"
This reverts commit 5e482ea9e5d8ec7f2a5448ea11fd41c04cb26a53.
