--- LeadKit/LeadKit/Classes/KeyboardNotificationValues/KeyboardDidNotificationValues.swift
@@ -1,17 +1,0 @@
-//
-//  KeyboardDidNotificationValues.swift
-//  LeadKit
-//
-//  Created by Ivan Smolin on 12/12/16.
-//  Copyright © 2016 Touch Instinct. All rights reserved.
-//
-
-import UIKit
-
-/// Struct which keeps values from one of those notifications:
-/// UIKeyboardDidShow, UIKeyboardDidHide, UIKeyboardDidChangeFrame
-public final class KeyboardDidNotificationValues: KeyboardNotificationValues {
-
-    // Animation key/value pairs are only available for the "will" family of notification. (UIWindow.h)
-    
-}
